moved marker styling into LocationMarkers
